refactor: 코드 정리 - 패키지명 변경 - 미사용 import 제거 - 미사용 테스트 코드 삭제
