+++ src/test/java/racingcar/domain/CarTest.java
@@ -1,11 +1,7 @@
-package racingcar.race;
+package racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.Car;
-import racingcar.domain.CarName;
-import racingcar.domain.Position;
-import racingcar.domain.RaceRecord;
 
 import static org.assertj.core.api.Assertions.assertThat;
