+++ src/test/java/racingcar/domain/CarsTest.java
@@ -3,9 +3,6 @@ package racingcar.domain;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.Car;
-import racingcar.domain.Position;
-import racingcar.race.Cars;
 import racingcar.rule.MinimumRule;
 import racingcar.rule.Rule;
 import racingcar.util.BasicNumberGenerator;
@@ -17,7 +14,7 @@ class CarsTest {
 
     @Test
     @DisplayName("자동차들을 이동시킨다. 이동된 자동차는 이동거리가 증가한다")
-    public void cars_move(){
+    public void cars_move() {
         Rule rule = new MinimumRule(4, new BasicNumberGenerator(4));
 
         List<Car> list = Arrays.asList(
