+++ src/test/java/racingcar/domain/RaceTest.java
@@ -1,9 +1,7 @@
-package racingcar.race;
+package racingcar.domain;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import racingcar.domain.Car;
-import racingcar.domain.Position;
 import racingcar.rule.MinimumRule;
 import racingcar.util.BasicNumberGenerator;
 
@@ -15,7 +13,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 class RaceTest {
     @ParameterizedTest(name = "입력값({0})이 경기 규칙 최소값({1}) 이상이면 자동차의 이동거리({2})가 증가한다")
     @CsvSource(value = {"3:4:0", "4:4:1", "5:4:1"}, delimiter = ':')
-    public void race_start_car_distacne_increase(int input, int minimum, int distance){
+    public void race_start_car_distance_increase(int input, int minimum, int distance) {
         List<Car> cars = Arrays.asList(
                 new Car("carA"),
                 new Car("carB"),
