--- src/test/java/study/ListTest.java
@@ -1,15 +1,0 @@
-package study;
-
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class ListTest {
-
-    @Test
-    public void test() {
-//        List<Car> cars = new ArrayList<>(null);
-//        cars.add(new Car());
-//        assertThat(cars.size()).isEqualTo(1);
-    }
-}
